+++ .idea/modules.xml
@@ -4,7 +4,6 @@
     <modules>
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" group="mobile/app" />
       <module fileurl="file://$PROJECT_DIR$/mobile.iml" filepath="$PROJECT_DIR$/mobile.iml" group="mobile" />
-      <module fileurl="file://$PROJECT_DIR$/app/mobile-app.iml" filepath="$PROJECT_DIR$/app/mobile-app.iml" group="mobile/app" />
     </modules>
   </component>
 </project>
