+++ app/src/main/res/layout/activity_select.xml
@@ -16,6 +16,18 @@
         android:text="사칙연산"
         android:textColor="#ffffff"
         android:textSize="20sp"/>
+    <Button
+        android:id="@+id/btn_tts"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:background="#000000"
+        android:layout_marginTop="50dp"
+        android:padding="20dp"
+        android:layout_gravity="center"
+        android:onClick="voice"
+        android:text="음성 문제"
+        android:textColor="#ffffff"
+        android:textSize="20sp"/>
 
     <Button
         android:id="@+id/btn_ranking"
@@ -39,6 +51,7 @@
         android:padding="20dp"
         android:layout_gravity="center"
         android:text="SETTING"
+        android:onClick="setting"
         android:textColor="#ffffff"
         android:textSize="20sp"/>
