+++ app/src/main/res/layout/activity_signup.xml
@@ -1,7 +1,10 @@
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
+    android:background="#ABCDEF"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    >
 <TextView
     android:id="@+id/WELCOME"
     android:textColor="#ffffff"
@@ -109,4 +112,4 @@
     app:layout_constraintHorizontal_bias="0.094"
     app:layout_constraintStart_toStartOf="parent"
     app:layout_constraintTop_toTopOf="parent" />
-</android.support.constraint.ConstraintLayout>
+</androidx.constraintlayout.widget.ConstraintLayout>
