+++ app/src/main/res/values/strings.xml
@@ -12,4 +12,8 @@
     <string name="action_submit">SUBMIT</string>
     <string name="NULL_MESSAGE">빈칸을 채워 모두 주세요</string>
     <string name="NOT_MATCH_MESSAGE">아이디 또는 패스워드가 올바르지 않습니다.</string>
+    <string name="setting_back">배경음악 음소거</string>
+    <string name="setting_effect">효과음 음소거</string>
+    <string name="voice_question1">아이스크림은 역시 민트초코</string>
+    <string name="voice_question2">1+1은 무엇일까요? 음성으로 답변해 주세요</string>
 </resources>
